Assignment Commit
Please feel free ignore this pull request. My class instructor wanted us to perform a test on some large open sourced system so I picked chromium. However, I still have a lot to learn. I wrote this additional test which I think it makes sure the two sequentially generated random bytes are not exactly the same. I am not exactly sure if the random is also strong enough that every time the sequence is different after initiation, or other ways to validate randomness would work here.

The likelihood of two random byte stream to be the same should be half of what the original code is testing assuming a 0.5 chance of either bits, so the edge case of this additional test should be even less frequent to fail. But as I said I couldn't even find where 'base::WriteInto' is at this moment so this might ruin the test too.